working with abstracted plugin
