--- _disable__vite-cep-plugin/.gitignore
@@ -1,1 +1,0 @@
-*.p12
