--- _disable__vite-cep-plugin/lib/node-built-ins.ts
@@ -1,30 +1,0 @@
-export const nodeBuiltIns = [
-  "crypto",
-  "assert",
-  "buffer",
-  "child_process",
-  "cluster",
-  "dgram",
-  "dns",
-  "domain",
-  "events",
-  "fs",
-  "http",
-  "https",
-  "net",
-  "os",
-  "path",
-  "punycode",
-  "querystring",
-  "readline",
-  "stream",
-  "string_decoder",
-  "timers",
-  "tls",
-  "tty",
-  "url",
-  "util",
-  "v8",
-  "vm",
-  "zlib",
-];
